+++ crates/libs/rbdc-mysql/src/types/timestamp.rs
@@ -13,7 +13,7 @@ use crate::value::MySqlValueFormat;
 
 impl Encode for Timestamp {
     fn encode(self, buf: &mut Vec<u8>) -> Result<usize, Error> {
-        let datetime = fastdate::DateTime::from_timestamp_millis(self.0 as i64);
+        let datetime = fastdate::DateTime::from_timestamp_millis(self.0);
         let size = date_time_size_hint(
             datetime.hour(),
             datetime.minute(),
