+++ crates/libs/rbdc-sqlite/src/statement/virtual.rs
@@ -177,7 +177,7 @@ fn prepare(
                 conn,
                 query_ptr,
                 query_len,
-                flags as u32,
+                flags,
                 &mut statement_handle,
                 &mut tail,
             )
